requirements.txt
asgiref==3.7.1
dj-database-url==2.0.0
Django==4.2.1
django-heroku==0.3.1
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
whitenoise==6.4.0
